# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Urlytix API
  version: 0.1.0
  description: |
    Contract for implemented and planned Urlytix endpoints. Operations marked
    `x-urlytix-status: planned` are design targets and are not available yet.
    Billing is retained for a later phase and its implemented routes return 404
    unless the server is explicitly started with `BILLING_ENABLED=true`.
    Registering a user atomically provisions an OWNER default workspace.
    Browser sessions use API-managed httpOnly `access_token` (15 minutes) and
    rotating `refresh_token` (30 days) cookies; bearer JWTs remain supported.
    All paths are relative to the NestJS `/api` global prefix.
servers:
  - url: https://api.example.invalid/api
    description: Replace with the environment API origin
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register a user
      description: Creates the user, an OWNER default workspace, and a refresh session.
      x-urlytix-status: implemented
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Authenticate a user
      x-urlytix-status: implemented
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "201":
          description: Authentication succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get the current user
      x-urlytix-status: implemented
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      summary: Rotate a refresh token and issue a new access token
      x-urlytix-status: implemented
      security:
        - refreshCookie: []
      responses:
        "201":
          description: Session refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Revoke the active refresh-token family
      x-urlytix-status: implemented
      security:
        - refreshCookie: []
      responses:
        "201":
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/profile:
    patch:
      tags: [Settings]
      operationId: updateProfile
      summary: Update current user profile
      x-urlytix-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: [string, "null"]
                  maxLength: 80
                email:
                  type: string
                  format: email
                timezone:
                  type: string
                  maxLength: 50
                locale:
                  type: string
                  maxLength: 10
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/forgot-password:
    post:
      tags: [Auth]
      operationId: forgotPassword
      summary: Request a password reset email
      x-urlytix-status: implemented
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "201":
          description: Generic acknowledgement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Reset password with emailed token
      x-urlytix-status: implemented
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 12
                  maxLength: 128
      responses:
        "201":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/change-password:
    post:
      tags: [Auth]
      operationId: changePassword
      summary: Change password and revoke other sessions
      x-urlytix-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 12
                  maxLength: 128
      responses:
        "201":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/sessions:
    get:
      tags: [Auth]
      operationId: listSessions
      summary: List active refresh sessions
      x-urlytix-status: implemented
      responses:
        "200":
          description: Active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [Auth]
      operationId: revokeOtherSessions
      summary: Revoke all sessions except the current one
      x-urlytix-status: implemented
      responses:
        "200":
          description: Other sessions revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/sessions/{sessionId}:
    delete:
      tags: [Auth]
      operationId: revokeSession
      summary: Revoke a single session
      x-urlytix-status: implemented
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /auth/2fa/setup:
    post:
      tags: [Auth]
      operationId: setupTwoFactor
      summary: Begin TOTP setup
      x-urlytix-status: implemented
      responses:
        "201":
          description: Setup payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [secret, otpauthUrl, qrCodeDataUrl]
                properties:
                  secret:
                    type: string
                  otpauthUrl:
                    type: string
                  qrCodeDataUrl:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/2fa/enable:
    post:
      tags: [Auth]
      operationId: enableTwoFactor
      summary: Enable TOTP after verifying a code
      x-urlytix-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code:
                  type: string
                  pattern: "^\\d{6}$"
      responses:
        "201":
          description: Enabled with one-time backup codes
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [message, backupCodes]
                properties:
                  message:
                    type: string
                  backupCodes:
                    type: array
                    items:
                      type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/2fa/disable:
    post:
      tags: [Auth]
      operationId: disableTwoFactor
      summary: Disable TOTP with password and code
      x-urlytix-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password, code]
              properties:
                password:
                  type: string
                code:
                  type: string
      responses:
        "201":
          description: Disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/2fa/verify:
    post:
      tags: [Auth]
      operationId: verifyTwoFactor
      summary: Complete login with TOTP or backup code
      x-urlytix-status: implemented
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [twoFactorToken, code]
              properties:
                twoFactorToken:
                  type: string
                code:
                  type: string
      responses:
        "201":
          description: Authentication succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource state conflicts with the request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          minimum: 400
          maximum: 599
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
    Message:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
          maxLength: 128
          description: Must include lowercase, uppercase, numeric, and symbol characters.
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    AuthResponse:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message:
          type: string
        accessToken:
          type: string
        requiresTwoFactor:
          type: boolean
        twoFactorToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
    AuthSession:
      type: object
      additionalProperties: false
      required: [id, createdAt, expiresAt, current]
      properties:
        id:
          type: string
          format: uuid
        userAgent:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        current:
          type: boolean
    TokenResponse:
      type: object
      additionalProperties: false
      required: [accessToken]
      properties:
        accessToken:
          type: string
    User:
      type: object
      additionalProperties: false
      required: [id, email, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: [string, "null"]
        email:
          type: string
          format: email
        timezone:
          type: string
        locale:
          type: string
        totpEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
  - accessCookie: []